Refactored to remove redundant files
