+++ src/lattice.py
